Move and rename `webview.ts` files
Fixes #165889
